+++ src/fruit/stand/people/Cashier.java
@@ -1,7 +1,7 @@
 package fruit.stand.people;
 
 public class Cashier extends Person {
-    public Cashier(String name, double cash) {
+    public Cashier(String name, int cash) {
         super(name, cash);
     }
 }
