+++ src/fruit/stand/people/Farmer.java
@@ -1,7 +1,7 @@
 package fruit.stand.people;
 
 public class Farmer extends Person {
-    public Farmer(String name, double cash) {
+    public Farmer(String name, int cash) {
         super(name, cash);
     }
 }
