+++ src/fruit/stand/products/Fruit.java
@@ -6,7 +6,7 @@ public class Fruit extends Product {
     //    private List<String> typeList;
     private String color;
 
-    protected Fruit(String name, String subtype, LocalDate expDate, double cost, int amount, String color) {
+    protected Fruit(String name, String subtype, LocalDate expDate, int cost, int amount, String color) {
         super(name, subtype, expDate, cost, amount);
         this.color = color;
     }
